requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.8
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.0
numpy==1.19.1
oauthlib==3.1.0
Pillow==7.0.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.8
Werkzeug==1.0.1
WTForms==2.3.3
